requirements.txt
appnope==0.1.0
asynctest==0.12.2
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
cursor==1.2.0
decorator==4.3.0
dill==0.2.8.2
docopt==0.6.2
enum34==1.1.6
Flask==1.0.2
genie==3.0.0
genie.abstract==3.0.1
genie.conf==3.0.1
genie.examples==3.0.1
genie.harness==3.0.4
genie.libs.conf==3.0.4
genie.libs.filetransferutils==3.0.1
genie.libs.ops==3.0.4
genie.libs.parser==3.0.7
genie.libs.robot==3.0.7
genie.libs.sdk==3.0.3
genie.libs.telemetry==3.0.3
genie.metaparser==3.0.1
genie.ops==3.0.1
genie.parsergen==3.0.2
genie.predcore==3.0.1
genie.telemetry==3.0.2
genie.utils==3.0.1
graphviz==0.8.3
gunicorn==19.9.0
halo==0.0.16
httplib2==0.11.3
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
ItsDangerous==1.0.0
jedi==0.12.1
Jinja2==2.10
jsonpickle==0.9.6
junit-xml==1.8
log-symbols==0.0.11
lxml==4.2.4
MarkupSafe==1.0
netaddr==0.7.19
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==5.4.6
ptyprocess==0.6.0
pyats==4.1.0
pyats.aereport==4.1.0
pyats.aetest==4.1.3
pyats.async==4.1.0
pyats.connections==4.1.0
pyats.datastructures==4.1.0
pyats.easypy==4.1.7
pyats.examples==4.1.0
pyats.kleenex==4.1.0
pyats.log==4.1.0
pyats.results==4.1.0
pyats.robot==4.1.1
pyats.tcl==4.1.1
pyats.templates==4.1.0
pyats.topology==4.1.2
pyats.utils==4.1.4
Pygments==2.2.0
python-dateutil==2.7.3
PyYAML==3.12
requests==2.19.1
robotframework==3.0.4
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.11.0
spinners==0.0.23
tabulate==0.8.2
termcolor==1.1.0
traitlets==4.3.2
unicon==3.1.2
urllib3==1.23
virlutils==0.8.4
wcwidth==0.1.7
Werkzeug==0.14.1
xmltodict==0.11.0
